// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        FarmerAdmin: {
            readonly id: number;
            full_name: string;
            sex: components["schemas"]["FarmerAdminSexEnum"];
            phone?: string | null;
            language?: string;
            reach_channel?: string;
            village: number;
            readonly village_name: string;
            readonly parish_name: string;
            readonly district_name: string;
            readonly registered_by_name: string | null;
            readonly user_phone: string | null;
        };
// ...
        VillageAdmin: {
            readonly id: number;
            parish: number;
            readonly parish_name: string;
            name: string;
        };
        VillageList: {
            id: number;
            name: string;
            parish: string;
            district: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}